Company Subscription Entities and Functionalities are created.

+++ backend/advantage/src/main/java/com/advantage/advantage/services/SingleAnalysisAdReportServiceImplementation.java
@@ -24,6 +24,9 @@ public class SingleAnalysisAdReportServiceImplementation implements SingleAnalys
     @Autowired
     TextualAdvertisementRepo adRepo;
 
+    @Autowired
+    TextualAdvertisementService textualAdvertisementService;
+
     private IgnoredPropertyCreator ignoredPropertyCreator;
     @Override
     public List<SingleAdAnalysisReport> getAllReports() {
@@ -114,8 +117,13 @@ public class SingleAnalysisAdReportServiceImplementation implements SingleAnalys
     }
 
     @Override
-    public SingleAdAnalysisReport deleteReportByReportId(long reportId) {
-        return reportRepo.deleteByReportId(reportId);
+    public boolean deleteReportByReportId(long reportId) {
+        List<SingleAdAnalysisReport> reports = reportRepo.findByReportId(reportId);
+        if(reports == null || reports.isEmpty()){
+            return false;
+        }
+        reportRepo.deleteByReportId(reportId);
+        return true;
     }
 }
